+++ src/pymmcore_widgets/_group_preset_widget/_edit_preset_widget.py
@@ -21,8 +21,8 @@ from qtpy.QtWidgets import (
 )
 from superqt.utils import signals_blocked
 
-from pymmcore_widgets._property_widget import PropertyWidget
 from pymmcore_widgets._util import block_core
+from pymmcore_widgets.device_properties._property_widget import PropertyWidget
 
 
 class EditPresetWidget(QDialog):
