+++ src/pymmcore_widgets/config_presets/__init__.py
@@ -1,0 +1,1 @@
+"""Widgets related to configuration groups and presets."""
