+++ src/pymmcore_widgets/device_properties/__init__.py
@@ -1,0 +1,1 @@
+"""Widgets related to device properties."""
